% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_gen.R
\name{dat_gen}
\alias{dat_gen}
\title{Function to generate data in Wyss et al 2025}
\usage{
dat_gen(n, scenario, seed1, seed2)
}
\arguments{
\item{n}{the number of observatoins to generate (size of data)}

\item{scenario}{a value of either 1 or 2. scenario=1 is the high-dimensional setting in Wyss et al 2025; scenario=2, observational study with 10 covariates in Ertefaie et al 2023}

\item{seed1}{seed for simulating random data}

\item{seed2}{seed for setting global parameters for scenario 1}
}
\value{
A dataframe of simulated data.
}
\description{
Function to generate data in Wyss et al 2025
}
\details{
Function to generate data used in Wyss et al. 2025
}
