% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal_model.R
\name{hal_model}
\alias{hal_model}
\title{HAL function from the package hal9001}
\usage{
hal_model(X, Y, max_degree, num_knots, nfolds, foldid)
}
\value{
A design matrix representing the expanded set of binary indicator functions of the covariates
}
\description{
HAL function from the package hal9001
}
\details{
The hal_model() function simply calls the fit_hal() function from the hal9001 R package and only returns the matrix of binary indicator basis function that is generated by the Highly Adaptive Lasso algorithm.
}
